refactor(app): extract command generation logic
